changed creator to admin

+++ rso_manage/views.py
@@ -4,22 +4,22 @@ from django.contrib import messages
 from django.urls import reverse
 from django.db import connection
 from .forms import RSOCreationForm
-from .models import RSO
-from .models import Registrations
+from .models import RSO, Registrations, RSOAdmin
 from users.models import Member
 
 def AddRSO(request):
     if request.method == 'POST':
         form = RSOCreationForm(request.POST, request.FILES)
-        if form.is_valid():
+        if form.is_valid() and not RSO.objects.filter(name=form.cleaned_data.get('name')).exists():
             name = form.cleaned_data.get('name')
             date_established = form.cleaned_data.get('date_established')
             college_association = form.cleaned_data.get('college_association')
             icon = form.cleaned_data.get('icon')
             description = form.cleaned_data.get('description')
             form_save = form.save(commit=False)
-            form_save.creator = request.user.username
             form.save()
+            new_rso_admin = RSOAdmin(member=Member.objects.get(username=request.user.username), rso=RSO.objects.get(name=name))
+            new_rso_admin.save()
             return redirect('home')
     else:
         form = RSOCreationForm()
@@ -37,7 +37,10 @@ def rso_profile(request, rso_name):
     member_names = [m.member.username for m in member_registrations]
     member_names = list(set(member_names))
     # print("names "+str(member_names))
-    return render(request, 'rso_profile.html', {'rso' : rso, 'member_registrations' : member_registrations, 'member_names' : member_names})
+    admin_registrations = RSOAdmin.objects.raw('SELECT * FROM "rso_manage_rsoadmin" WHERE rso_id = {}'.format(rso_id))
+    admin_names = list(set([m.member.username for m in admin_registrations]))
+    return render(request, 'rso_profile.html', {'rso' : rso, 'member_registrations' : member_registrations, 'member_names' : member_names,
+                                                'admin_registrations' : admin_registrations, 'admin_names' : admin_names})
 
 def register(request, rso_name):
     username = request.user.username
@@ -49,28 +52,33 @@ def register(request, rso_name):
     return redirect('/rsos/'+rso_name+'/profile')
 
 def unregister(request, rso_name):
-    username = request.user.username
-    member = Member.objects.get(username=username)
+    member = Member.objects.get(username=request.user.username)
     rso = RSO.objects.get(name=rso_name)
+
     if Registrations.objects.filter(member=member, rso=rso).exists():
         Registrations.objects.get(member=member, rso=rso).delete()
     return redirect('/rsos/'+rso_name+'/profile')
-    return render(request, 'rso_list.html', {'all_rsos' : all_rsos})
 
 def rso_delete(request, rso_name):
     rso_id = RSO.objects.get(name=rso_name).id
-    if request.user.is_authenticated:
+    admin_registrations = RSOAdmin.objects.raw('SELECT * FROM "rso_manage_rsoadmin" WHERE rso_id = {}'.format(rso_id))
+    admin_names = list(set([m.member.username for m in admin_registrations]))
+    if request.user.is_authenticated and request.user.username in admin_names:
         try:
             cursor = connection.cursor()
             query = 'DELETE FROM "events_event" WHERE rso_id = {}'.format(rso_id)
             cursor.execute(query)
-            query = 'DELETE FROM "rso_manage_rso" WHERE rso_manage_rso.name = "{}" AND rso_manage_rso.creator = "{}"'.format(rso_name, request.user.username)
+            query = 'DELETE FROM "rso_manage_rso" WHERE rso_manage_rso.name = "{}"'.format(rso_name)
+            cursor.execute(query)
+            query = 'DELETE FROM "rso_manage_rsoadmin" WHERE rso_manage_rsoadmin.name = "{}"'.format(rso_name)
+            cursor.execute(query)
+            query = 'DELETE FROM "rso_manage_registrations" WHERE rso_manage_registrations.name = "{}"'.format(rso_name)
             cursor.execute(query)
             connection.commit()
             messages.success(request, "RSO deleted")
         except Exception as E:
             print(E)
-            messages.error(request, "RSO not found or you must be the creator of the RSO")
+            messages.error(request, "RSO not found or you must be an admin of the RSO")
     else:
         messages.error(request, "You must be logged in to delete an RSO.")
     return redirect('/rsos/')
